Extract large-file chunked loading into engine/src/largeFileChunking.ts
Moves handleStartLargeFile, handleLargeFileChunk, and
handleLargeFileComplete out of main.ts as thin wrappers.
